WIP - start asymptote
